[Nvidia] [Smartswitch] Added check for file while checking for reboot cause (#25924)
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
Confirm that the PCI tree exists before checking for the watchdog reboot cause, as this reboot cause is dependent on the reading register and the pci device should be available for this to be done, otherwise there are error lgos

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
check for the path if it exists using os.path.exists function

#### How to verify it

`get_reboot_cause` execution while the DPU is powered off, we should not get the following error log:
`ERR pmon#chassisd[38]: Failed to check watchdog reason for DPU0! Command '['mlxreg', '-d', '0000:08:00.0', '--reg_name', 'MRSI', '-g', '-indexes', 'device=1']' returned non-zero exit status 1.`

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505
- [ ] 202511

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->


#### A picture of a cute animal (not mandatory but encouraged)

+++ platform/mellanox/mlnx-platform-api/sonic_platform/module.py
@@ -1,6 +1,6 @@
 #
 # SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
-# Copyright (c) 2021-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# Copyright (c) 2021-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -21,8 +21,9 @@ import threading
 from sonic_platform_base.module_base import ModuleBase
 from sonic_platform_base.chassis_base import ChassisBase
 from sonic_py_common.syslogger import SysLogger
-from .dpuctlplat import DpuCtlPlat, BootProgEnum
+from .dpuctlplat import DpuCtlPlat, BootProgEnum, PCI_DEV_BASE
 import subprocess
+import os
 
 from . import utils
 from .device_data import DeviceDataManager, DpuInterfaceEnum
@@ -438,22 +439,24 @@ class DpuModule(ModuleBase):
             REBOOT_CAUSE_DPU_SELF_REBOOT
         """
         # Check for Watchdog reboot first
-        pcie_path = DeviceDataManager.get_dpu_interface(self._name.lower(), DpuInterfaceEnum.PCIE_INT.value)
-        # mlxreg -d 0000:08:00.0 --reg_name MRSI -g -indexes "device=1"
-        try:
-            op = subprocess.check_output(['mlxreg', '-d', pcie_path, '--reg_name', 'MRSI', '-g', '-indexes', 'device=1'])
-        except subprocess.CalledProcessError as e:
-            logger.log_error(f"Failed to check watchdog reason for {self._name}! {e}")
-            op = b''
-        reset_reason_value = None
-        for line in op.decode().split('\n'):
-            if 'reset_reason' in line:
-                # Extract the value after the '|'
-                reset_reason_value = line.split('|')[1].strip()
-                break
-        if reset_reason_value and int(reset_reason_value,16) == self.MLX_DPU_REBOOT_CAUSE_WATCHDOG:
-            logger.log_notice(f"Reset reason for {self._name} is {ChassisBase.REBOOT_CAUSE_WATCHDOG}")
-            return ChassisBase.REBOOT_CAUSE_WATCHDOG, 'Watchdog reboot'
+        pcie_dev_id = DeviceDataManager.get_dpu_interface(self._name.lower(), DpuInterfaceEnum.PCIE_INT.value)
+        pcie_path = os.path.join(PCI_DEV_BASE, pcie_dev_id)
+        if os.path.exists(pcie_path):
+            try:
+                # mlxreg -d 0000:08:00.0 --reg_name MRSI -g -indexes "device=1"
+                op = subprocess.check_output(['mlxreg', '-d', pcie_dev_id, '--reg_name', 'MRSI', '-g', '-indexes', 'device=1'])
+            except subprocess.CalledProcessError as e:
+                logger.log_error(f"Failed to check watchdog reason for {self._name}! {e}")
+                op = b''
+            reset_reason_value = None
+            for line in op.decode().split('\n'):
+                if 'reset_reason' in line:
+                    # Extract the value after the '|'
+                    reset_reason_value = line.split('|')[1].strip()
+                    break
+            if reset_reason_value and int(reset_reason_value,16) == self.MLX_DPU_REBOOT_CAUSE_WATCHDOG:
+                logger.log_notice(f"Reset reason for {self._name} is {ChassisBase.REBOOT_CAUSE_WATCHDOG}")
+                return ChassisBase.REBOOT_CAUSE_WATCHDOG, 'Watchdog reboot'
         # Check for other reboot causes
         for f, rd in self.reboot_cause_map.items():
             if utils.read_int_from_file(f) == 1:
